Reworked the authoring using annotations.
Complete reworking of the authoring tools.  Trying to work in a
more declarative manner, using annotations instead of procedural method
calls.  See the org/sc/obo/annotations/ProteinSite.java class
for an example of how the new declarations work, and see the main
method in OBOAnnotationParser for an example of some of the methods you
can call on these classes.