Added @Term creation from OBOTerm objects.
Updated OBOTermCreator to create @Term-annotated Class objects from
OBOTerm objects derived from an OBOOntology.  The trick here is that we
have to *recursively* create terms for the super-classes as well, and
(finally) we don't want to repeatedly do that -- so we cache the previously
created terms by ID, which should be unique.