Better Java code generation.
Fixed a variety of bugs, and added several new major features,
to the Java class creation in OBOTermCreator and in the Java code
generation in JavaWriter and JavaExporter.

The main features added were (a) better Java name handling, so that
just translating the names of the terms into java didn't cause a
thousand errors, and (b) better (well, longer but non-conflicting)
method naming in OBOTermCreator.  Also, correctly handled the
annotations that were added to each class file, as well as to the
relation methods.

See the comments in OBOTermCreator.java for more details.